worker/requirements.txt
celery
redis
neo4j
psycopg2-binary
pydantic
python-dotenv
httpx
